start implementing the login view in mvp style
