+++ teamgreen-pms-app/src/main/java/ch/bfh/bti7081/s2017/green/ui/components/login/LoginView.java
@@ -5,5 +5,5 @@ import com.vaadin.navigator.View;
 public interface LoginView extends View {
     void addListener(LoginViewListener listener);
 
-    void login();
+    void onAfterBeanInitializaiton ();
 }
