+++ teamgreen-pms-app/src/main/java/ch/bfh/bti7081/s2017/green/ui/components/login/LoginViewPresenter.java
@@ -18,7 +18,5 @@ public class LoginViewPresenter implements LoginViewListener {
 
     @Override
     public void onLoginClick() {
-        loginView.login();
-
     }
 }
